third_party/blink/web_tests/fast/shapes/shape-outside-floats/shape-outside-floats-image-margin-001-expected.txt

This test requires the Ahem font. The green content should wrap around the blue rounded rectangle.

X

X
X

X
XXXX
PASS successfullyParsed is true

TEST COMPLETE

PASS imageShapeRect('a').top is 0
PASS imageShapeRect('b').top is 50
PASS imageShapeRect('b').left is 300
PASS imageShapeRect('c').top is 75
PASS imageShapeRect('c').left is 300
PASS imageShapeRect('d').top is 125
PASS imageShapeRect('e').top is 150
PASS imageShapeRect('e').left is 0
